code/Backend_Server/Finance-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pydantic==2.6.4
python-jose==3.3.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pytest==8.1.1
httpx==0.27.0
